Create README.md
I will  be working to understand the results of an A/B test run by an e-commerce website. The company has developed a new web page in order to try and increase the number of users who "convert," meaning the number of users who decide to pay for the company's product. we'll help the company understand if they should implement this new page, keep the old page, or perhaps run the experiment longer to make their decision.